purge adbd entirely, its only used on Ubuntu, and maintained there
The Ubuntu team includes adbd to run on Ubuntu Touch on mobile devices.
adbd should never be installed on Debian, its very insecure.  The Ubuntu
team has been maintaining adbd in their own package, so there is no need to
keep the out-of-date versions of adbd packaging in the Debian package.